+++ packages/perseus-editor/src/all-editors.ts
@@ -13,7 +13,7 @@ import IframeEditor from "./widgets/iframe-editor";
 import ImageEditor from "./widgets/image-editor";
 import InputNumberEditor from "./widgets/input-number-editor";
 import InteractionEditor from "./widgets/interaction-editor";
-import InteractiveGraphEditor from "./widgets/interactive-graph-editor";
+import InteractiveGraphEditor from "./widgets/interactive-graph-editor/interactive-graph-editor";
 import LabelImageEditor from "./widgets/label-image-editor";
 import MatcherEditor from "./widgets/matcher-editor";
 import MatrixEditor from "./widgets/matrix-editor";